fix: remove command palette
