+++ package.json
@@ -11,10 +11,9 @@
     "prepare": "husky",
     "build:manifest:chrome": "mv dist/manifestv3.json dist/manifest.json",
     "build:background:chrome": "vite build --config vite.chrome.background.config.ts",
-    "build:js": "vite build --config vite.content.config.ts",
     "build:web": "tsc && vite build",
-    "build:firefox": "cross-env NODE_ENV=production run-s build:js build:web",
-    "build:chrome": "cross-env NODE_ENV=production run-s build:js build:background:chrome build:web build:manifest:chrome",
+    "build:firefox": "cross-env NODE_ENV=production run-s build:web",
+    "build:chrome": "cross-env NODE_ENV=production run-s build:background:chrome build:web build:manifest:chrome",
     "sign:firefox": "web-ext sign --use-submission-api --channel unlisted --source-dir firefox-dist/ --artifacts-dir .",
     "publish:chrome": "chrome-webstore-upload upload --source taby.zip --auto-publish",
     "publish:firefox": "web-ext sign --channel listed --source-dir firefox-dist/"
