+++ public/manifest.json
@@ -10,14 +10,6 @@
     "default_popup": "./src/popup/popup.html",
     "default_title": "Open the popup"
   },
-  "content_scripts": [
-    {
-      "matches": ["<all_urls>"],
-      "run_at": "document_end",
-      "js": ["./index.global.js"],
-      "css": ["./taby.css"]
-    }
-  ],
   "background": {
     "page": "src/background/background.html",
     "persistent": false
@@ -30,12 +22,6 @@
       "suggested_key": {
         "default": "Alt+Q"
       }
-    },
-    "TOGGLE_MENU": {
-      "suggested_key": {
-        "default": "Alt+W"
-      },
-      "description": "Toggle Taby menu"
     },
     "DUPLICATE_TAB": {
       "description": "Duplicate current tab"
