--- src/content-scripts/content-script.css
@@ -1,13 +1,0 @@
-.taby-menu {
-  position: fixed;
-  z-index: 100000000;
-  border-radius: 6px !important;
-}
-
-.taby-searchList {
-  scrollbar-width: none !important;
-}
-
-.taby-searchList::-webkit-scrollbar {
-  display: none !important;
-}
