+++ src/lib/window.ts
@@ -1,36 +1,11 @@
-import { EContext, EStorage } from "../type/misc.ts";
-import { MutableRefObject } from "react";
-import browser from "webextension-polyfill";
-
-const SEARCH_INPUT_SIZE: number = 55,
-  BORDER_SIZE: number = 2,
-  PADDINGS_SEARCH_LIST: number = 16,
-  SEARCH_ITEM_SIZE: number = 33,
-  NAV_POPUP: number = 30;
-
-export const computeWindowSize = async (context: EContext) => {
-  const storage = await browser.storage.local.get([EStorage.PositionBlock]);
-
-  const position_block = 1 - Number(storage[EStorage.PositionBlock]) * 0.01;
-
-  const window_size =
-    context === EContext.ContentScript
-      ? window.innerHeight * position_block
-      : 600 - NAV_POPUP;
-
-  const menu_size =
-    window_size - SEARCH_INPUT_SIZE - PADDINGS_SEARCH_LIST - BORDER_SIZE;
-  const capacity = Math.floor(menu_size / SEARCH_ITEM_SIZE);
-
-  return context == EContext.ContentScript ? Math.min(10, capacity) : capacity;
-};
+import { RefObject } from "react";
 
 export function moveUp(
   idx: number,
   size: number,
   capacity: number,
-  start: MutableRefObject<number>,
-  end: MutableRefObject<number>,
+  start: RefObject<number>,
+  end: RefObject<number>,
   setElement: (value: number) => void,
 ) {
   const next = (idx - 1 + size) % size;
@@ -49,8 +24,8 @@ export function moveDown(
   idx: number,
   size: number,
   capacity: number,
-  start: MutableRefObject<number>,
-  end: MutableRefObject<number>,
+  start: RefObject<number>,
+  end: RefObject<number>,
   setElement: (value: number) => void,
 ) {
   const next = (idx + 1) % size;
