+++ src/popup/popup.tsx
@@ -1,13 +1,11 @@
 import "../index.css";
-import "../preflight.css";
 import "./popup.css";
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import App from "~/components/App";
-import { EContext } from "~/type/misc";
 
 createRoot(document.getElementById("taby-root")!).render(
   <StrictMode>
-    <App context={EContext.Popup} />
+    <App />
   </StrictMode>,
 );
