--- src/preflight.css
@@ -1,1 +1,0 @@
-@import "tailwindcss/preflight.css" layer(base);
