--- vite.content.config.ts
@@ -1,31 +1,0 @@
-import react from "@vitejs/plugin-react-swc";
-import { resolve } from "path";
-import path from "path";
-import { defineConfig } from "vite";
-
-export default defineConfig({
-  plugins: [react()],
-  resolve: {
-    alias: {
-      "~": path.resolve(__dirname, "./src"),
-    },
-  },
-  define: {
-    "process.env": {},
-  },
-  build: {
-    emptyOutDir: true,
-    outDir: resolve(__dirname, "dist"),
-    lib: {
-      formats: ["iife"],
-      entry: resolve(__dirname, "./src/content-scripts/content-script.tsx"),
-      name: "Taby",
-    },
-    rollupOptions: {
-      output: {
-        entryFileNames: "index.global.js",
-        extend: true,
-      },
-    },
-  },
-});
